Factor out VideoMode
